add get by id

+++ core/src/main/java/io/fundrequest/core/request/infrastructure/RequestRepository.java
@@ -4,10 +4,9 @@ import io.fundrequest.core.infrastructure.repository.JpaRepository;
 import io.fundrequest.core.request.domain.Request;
 import org.springframework.data.jpa.repository.Query;
 
-import javax.swing.text.html.Option;
 import java.util.Optional;
 
-public interface RequestRepository extends JpaRepository<Request, Integer> {
+public interface RequestRepository extends JpaRepository<Request, Long> {
     @Query("SELECT r FROM Request r where r.issueInformation.link = ?1")
     Optional<Request> findByIssueLink(String link);
 }
